+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.2.1] -
+### Added
+- Raises an exception if `value` parameter isn't number nor expression.
 ### Changed
 - `hello-world` example uses `div` instead of `p`.
 ### Fixed